requirements.txt
fastapi~=0.110.0
torch==2.0.0
torchvision==0.15.1
opencv-python==4.6.0.66
uvicorn~=0.29.0
pillow~=10.2.0
pydantic~=2.6.4